refactor: pages to src
