+++ src/components/Card/Card.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, test } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { Card } from './Card';
-import { dataMock } from '@/test/mocks/dataMock';
+import { dataMock } from '../../test/mocks/dataMock';
 
 describe('Card', () => {
   test('ensure that the card component renders the relevant card data', async () => {
