+++ src/pages/globals.css
